+++ apps/meteor/server/services/omnichannel-analytics/service.ts
@@ -111,13 +111,13 @@ export class OmnichannelAnalyticsService extends ServiceClassInternal implements
 				const hour = parseInt(m.add(currentHour ? 1 : 0, 'hour').format('H'));
 				const label = {
 					from: moment.utc().set({ hour }).tz(timezone).format('hA'),
-					to: moment.utc().set({ hour }).add(1, 'hour').tz(timezone).format('hA'),
+					to: moment.utc().set({ hour }).endOf('hour').tz(timezone).format('hA'),
 				};
 				data.dataLabels.push(`${label.from}-${label.to}`);
 
 				const date = {
 					gte: m.toDate(),
-					lt: moment(m).add(1, 'hours').toDate(),
+					lte: moment(m).endOf('hour').toDate(),
 				};
 
 				data.dataPoints.push(await this.chart.callAction(chartLabel, date, departmentId, extraQuery));
@@ -128,7 +128,7 @@ export class OmnichannelAnalyticsService extends ServiceClassInternal implements
 
 				const date = {
 					gte: m.toDate(),
-					lt: moment(m).add(1, 'days').toDate(),
+					lte: moment(m).endOf('day').toDate(),
 				};
 
 				data.dataPoints.push(await this.chart.callAction(chartLabel, date, departmentId, extraQuery));